// src/app/features/Course/Models/course-details-dto.interface.ts
export interface ICourseDetailsDto {
    courseID: string;
    teacherID: string;
    teacherName: string;
    teacherTitle: string;
    teacherBio: string;
    teacherImgUrl: string;
    categoryName: string;
    title: string;
    subtitle: string;
    description: string;
    durationInSeconds: number;
    level: string;
    levelAr: string;
    thumbnailURL: string;
    promoVideoURL: string;
    rating: number;
    reviewsCount: number;
    price: number;
    language: string;
    languageAr: string;

}